Made the commands easier to use
Display the command list when the user uses the base command in addition
to /hd help. Move the plugin information to the subcommand /hd version
