+++ Plugin/src/main/java/com/gmail/filoghost/holographicdisplays/commands/main/subs/VersionCommand.java
@@ -1,0 +1,52 @@
+package com.gmail.filoghost.holographicdisplays.commands.main.subs;
+
+import java.util.Arrays;
+import java.util.List;
+
+import org.bukkit.command.CommandSender;
+
+import com.gmail.filoghost.holographicdisplays.HolographicDisplays;
+import com.gmail.filoghost.holographicdisplays.commands.Colors;
+import com.gmail.filoghost.holographicdisplays.commands.Strings;
+import com.gmail.filoghost.holographicdisplays.commands.main.HologramSubCommand;
+import com.gmail.filoghost.holographicdisplays.exception.CommandException;
+
+public class VersionCommand extends HologramSubCommand {
+	
+	public VersionCommand() {
+		super("version");
+		setPermission(Strings.BASE_PERM + "version");
+	}
+
+	@Override
+	public String getPossibleArguments() {
+		return "";
+	}
+
+	@Override
+	public int getMinimumArguments() {
+		return 0;
+	}
+
+
+	@Override
+	public void execute(CommandSender sender, String label, String[] args) throws CommandException {
+		sender.sendMessage("");
+		sender.sendMessage(Colors.PRIMARY_SHADOW + "Server is running " + Colors.PRIMARY + "Holographic Displays " + Colors.PRIMARY_SHADOW + "v" + HolographicDisplays.getInstance().getDescription().getVersion() + " by " + Colors.PRIMARY + "filoghost");
+		if (sender.hasPermission(Strings.BASE_PERM + "help")) {
+			sender.sendMessage(Colors.PRIMARY_SHADOW + "Commands: " + Colors.PRIMARY + "/" + label + " help");
+		}
+	}
+
+	@Override
+	public List<String> getTutorial() {
+		return Arrays.asList("Displays plugin information.");
+	}
+	
+	@Override
+	public SubCommandType getType() {
+		return SubCommandType.GENERIC;
+	}
+
+
+}
